refactor: move pipes to lin
