+++ libs/shared/ui/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/shared-ui/shared-ui';
+export * from './lib/pipes';
