+++ libs/shared/ui/src/lib/pipes/encode-uri.pipe.ts
