+++ libs/shared/ui/src/lib/pipes/exercise-difficulty.pipe.ts
